Lion Leftovers/Feedback Management Service/requirements.txt
beautifulsoup4==4.8.2
boto3==1.11.2
botocore==1.14.2
bs4==0.0.1
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
flasgger==0.9.4
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-RESTful==0.3.7
flask-restful-swagger==0.20.1
Flask-Track-Usage==2.0.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
tika==1.23.1
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
xlrd==1.2.0
psycopg2-binary==2.9.9
pandas==1.5.0
MarkupSafe==2.0.1